[feat] 로컬 캐시 업데이트를 Repository에서 하도록 변경
