+++ nottokeyword/src/main/java/com/example/nottokeyword/datasource/remote/KeywordRemoteDataSource.kt
@@ -1,10 +1,12 @@
 package com.example.nottokeyword.datasource.remote
 
+import com.example.nottokeyword.Keyword
+
 interface KeywordRemoteDataSource {
 
     suspend fun insertKeyword(content: String): Boolean
 
-    suspend fun getKeywords()
+    suspend fun getKeywords(callback: (List<Keyword>) -> Unit)
 
     suspend fun deleteKeyword(keyword: String): Boolean
 }
